+++ services/src/handlers/layers.rs
@@ -41,6 +41,8 @@ async fn list_root_collections_handler<C: Context>(
             },
             name: "Datasets".to_string(),
             description: "Basic Layers for all Datasets".to_string(),
+            entry_label: None,
+            properties: None,
         }));
 
         options.limit -= 1;
@@ -56,6 +58,8 @@ async fn list_root_collections_handler<C: Context>(
             },
             name: "Layers".to_string(),
             description: "All available Geo Engine layers".to_string(),
+            entry_label: None,
+            properties: None,
         }));
 
         options.limit -= 1;
@@ -82,6 +86,8 @@ async fn list_root_collections_handler<C: Context>(
             },
             name: provider_listing.name,
             description: provider_listing.description,
+            entry_label: None,
+            properties: None,
         }));
     }