+++ CHANGELOG.md
@@ -29,6 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Removed the custom Geo Bon EBV portal handlers. Instead, the EBV hierarchy is now browsed through the layer collection API.
+
+  - https://github.com/geo-engine/geoengine/pull/581
 
 - Changed the temporal reference of regular raster time series from a start date to a valid time interval `dataTime`. Before and after the `dataTime`, only one loading info with nodata will be produced instead of lots of nodata-chunks with the original interval of the time series.
 
@@ -40,7 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - https://github.com/geo-engine/geoengine/pull/561
   - `MaskedGrid` replaces `Grid` in `GridOrEmpty` / `RasterTile`
   - GeoTIFF files created by the engine contain the validity mask if not specified otherwise.
-  
+
 - Refactored dataset ids and external provders
 
   - https://github.com/geo-engine/geoengine/pull/554